trade related with Responsible

+++ src/main/java/com/esgi/group5/jeeproject/domain/use_cases/trades/DeleteTrade.java
@@ -1,7 +1,10 @@
 package com.esgi.group5.jeeproject.domain.use_cases.trades;
 
+import com.esgi.group5.jeeproject.domain.models.Trade;
 import com.esgi.group5.jeeproject.domain.models.User;
 import com.esgi.group5.jeeproject.domain.repositories.TradeRepository;
+
+import java.util.Optional;
 
 public class DeleteTrade {
     private final TradeRepository tradeRepository;
@@ -11,7 +14,11 @@ public class DeleteTrade {
     }
 
     public boolean deleteTrade(Long tradeId, User user) {
-        //TODO: Check if user is the responsible of the concerned Trade
+        Optional<Trade> trade = tradeRepository.getTradeById(tradeId);
+        if (trade.isEmpty())
+            return false;
+        if (trade.get().getResponsible().getId() != user.getId())
+            return false;
         tradeRepository.deleteTradeById(tradeId);
         return true;
     }
